backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Pillow==12.1.1
ffmpeg-python==0.2.0
Werkzeug==3.0.1